rag/requirements.txt
transformers==4.44.2
torch==2.4.1
tiktoken==0.7.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
openai==1.44.0
langchain==0.2.14
langchain-community==0.2.12
python-dotenv==1.0.1
langchain-openai==0.1.23 
langchain-chroma==0.1.3 
beautifulsoup4==4.12.3 